Fix use config file in the current workspace
I had tried using the exercism config file in the users home directory,
however this failed since it expects the exercism workspace to be in a
directory under the current user and not off the root. I have changed
this to mount a .config/exercism directory at the current directory and
added this directory to the gitignore file. This way each exercism track
is self contained.